Move some dict functions into utils
These will be re-used elsewhere in the code.
