+++ hexrd/ui/calibration/picks_tree_view_dialog.py
@@ -17,6 +17,7 @@ from hexrd.ui.hexrd_config import HexrdConfig
 from hexrd.ui.tree_views.picks_tree_view import PicksTreeView
 from hexrd.ui.ui_loader import UiLoader
 from hexrd.ui.utils.conversions import angles_to_cart, cart_to_angles
+from hexrd.ui.utils.dicts import ensure_all_keys_match, ndarrays_to_lists
 
 
 class PicksTreeViewDialog:
@@ -94,14 +95,7 @@ class PicksTreeViewDialog:
 
         # Our tree view is expecting lists rather than numpy arrays.
         # Go ahead and perform the conversion...
-        def ndarray_to_lists(d):
-            for k, v in d.items():
-                if isinstance(v, dict):
-                    ndarray_to_lists(v)
-                elif isinstance(v, np.ndarray):
-                    d[k] = v.tolist()
-
-        ndarray_to_lists(self.dictionary)
+        ndarrays_to_lists(self.dictionary)
         self.tree_view.model().config = self.dictionary
         self.tree_view.rebuild_tree()
         self.tree_view.expand_rows()
@@ -110,34 +104,15 @@ class PicksTreeViewDialog:
         # This will validate and sort the keys to match that of the
         # internal dict we already have.
         # All of the dict keys must match exactly.
-        def recurse(this, other, ret, path):
-            this_keys = sorted(this.keys())
-            other_keys = sorted(other.keys())
-            if this_keys != other_keys:
-                this_keys_str = ', '.join(f'"{x}"' for x in this_keys)
-                other_keys_str = ', '.join(f'"{x}"' for x in other_keys)
-                msg = (
-                    f'Current keys {this_keys_str} failed to match import '
-                    f'data keys {other_keys_str}'
-                )
-                if path:
-                    path_str = ' -> '.join(path)
-                    msg += f' for path "{path_str}"'
+        try:
+            ret = ensure_all_keys_match(self.dictionary, data)
+        except KeyError as e:
+            msg = e.args[0]
+            msg += f'\nin file "{filename}"\n'
+            msg += '\nPlease be sure the same settings are being used.'
 
-                msg += f' in file "{filename}"'
-                msg += '\n\nPlease be sure the same settings are being used.'
-                QMessageBox.critical(self.ui, 'HEXRD', msg)
-                raise Exception(msg)
-
-            for k, v in this.items():
-                if isinstance(v, dict):
-                    ret[k] = {}
-                    recurse(v, other[k], ret[k], path + [k])
-                else:
-                    ret[k] = other[k]
-
-        ret = {}
-        recurse(self.dictionary, data, ret, [])
+            QMessageBox.critical(self.ui, 'HEXRD', msg)
+            raise KeyError(msg)
 
         # Update the validated data
         data.clear()
